+++ ootrando_overworldmap_hamsda/scripts/init.lua
@@ -67,9 +67,9 @@ if has_map then
     Tracker:AddItems("items/capture_entrance.json")
     Tracker:AddItems("items/options_entrance.json")
   else
-    Tracker:AddItems("items/sequences.json")
     Tracker:AddItems("items/capture_spawns.json")
   end
+  Tracker:AddItems("items/sequences.json")
   Tracker:AddItems("items/capture_items.json")
   Tracker:AddItems("items/tricks.json")
   Tracker:AddItems("items/options.json")